Revert "Changing the examples to add support for docker"
